B #464: VM: add and share raw section
